+++ frontend/src/app/about/page.tsx
@@ -1,65 +1,26 @@
-'use client';
+import { Metadata } from 'next';
+import { images } from '@/constants';
+import AboutMain from '@/app/about/AboutMain';
 
-import Image from 'next/image';
-import update from '@/data/updates.json';
-import developer from '@/data/developer.json';
-import { useState } from 'react';
+export const metadata: Metadata = {
+  title: 'MetaPick(메타픽) - 정보',
+  description: '개발자 정보와 업데이트 내역을 확인하세요.',
+  openGraph: {
+    title: 'MetaPick(메타픽) - 정보',
+    description: '개발자 정보와 업데이트 내역을 확인하세요.',
+    images: [images.THUMBNAIL],
+    url: `${process.env.NEXT_PUBLIC_URL}/about`,
+    type: 'website',
+    siteName: 'MetaPick(메타픽)',
+  },
+};
 
-export default function About() {
-  const [openIndex, setOpenIndex] = useState<number | null>(0); // 어떤 항목 열려있는지
-
-  const handleToggle = (index: number) => {
-    setOpenIndex(openIndex === index ? null : index); // 다시 누르면 닫힘
-  };
-
+function About() {
   return (
-    <main className="flex flex-col items-center justify-center bg-mainBlack text-white space-y-12 mt-20">
-      {/* 공지사항 */}
-      <section className="w-3/4 max-w-2xl">
-        <h2 className="text-xl font-bold mb-4 border-b pb-2">☑️ 업데이트</h2>
-        <ul className="space-y-2">
-          {update.updates.map((update, index) => (
-            <li key={index} className="border-b border-gray-700 pb-2">
-              <button
-                onClick={() => handleToggle(index)}
-                className="flex justify-between w-full text-left hover:text-gray-400"
-              >
-                <span>{update.title}</span>
-                <span className="text-gray-400 text-sm">{update.date}</span>
-              </button>
-              {openIndex === index && <p className="mt-2 text-gray-300 text-sm">{update.content}</p>}
-            </li>
-          ))}
-        </ul>
-      </section>
-
-      {/* 개발자 소개 */}
-      <section className="w-3/4 max-w-2xl">
-        <h3 className="text-xl font-bold mb-4 border-b pb-2">🧑🏻‍💻 개발자</h3>
-        <div className="grid grid-cols-1 sm:grid-cols-3 gap-6">
-          {developer.developers.map((dev, index) => (
-            <div key={index} className="flex flex-col items-center p-4 border border-gray-700 rounded-lg">
-              <div className="w-16 h-16 bg-gray-600 rounded-full mb-2">
-                <Image alt="" src={dev.img} width={64} height={64} className="rounded-full object-cover" />
-              </div>
-              {/* 프로필 이미지 자리 */}
-              <h3 className="font-semibold">{dev.name}</h3>
-              <p className="text-gray-400">{dev.role}</p>
-              <a href={`mailto:${dev.email}`} className="text-blue-400 mt-2 text-sm hover:underline">
-                {dev.email}
-              </a>
-              <a
-                target="_blank"
-                rel="noopener noreferrer"
-                href={dev.github}
-                className="text-blue-400 mt-2 text-sm hover:underline"
-              >
-                Github
-              </a>
-            </div>
-          ))}
-        </div>
-      </section>
-    </main>
+    <>
+      <AboutMain />
+    </>
   );
 }
+
+export default About;
