+++ frontend/src/app/socketRoom/page.tsx
@@ -5,13 +5,15 @@ import BanpickSocket from '@/app/socketRoom/banpickSocket';
 import Loading from '@/components/Loading';
 
 export const metadata: Metadata = {
-  title: 'MetaPick - 대기방',
+  title: 'MetaPick(메타픽) - 대기방',
   description: '팀모드 대기방입니다.',
   openGraph: {
-    title: 'MetaPick - 대기방',
+    title: 'MetaPick(메타픽) - 대기방',
     description: '팀모드 대기방입니다.',
     images: [images.THUMBNAIL],
-    url: process.env.NEXT_PUBLIC_URL,
+    url: `${process.env.NEXT_PUBLIC_URL}/socketRoom`,
+    type: 'website',
+    siteName: 'MetaPick(메타픽)',
   },
 };
